Start to lay out the basic app framework.
Application is based on the "component" clojure library approach to modularization, and leverages environ to inject configuration. Database is roughly setup to connect to an OrientDB instance, with "sane default" of connecting to a local file instead of a server. The assumption is made that there will be a parallel build in grunt to generate html/js/css/etc for the site itself, but that is not yet included.