Add basic unit tests, fix errors they exposed.
There were some less than ideal situations exposed by basic unit testing of the static resource fetching. In particular, it was returning a status code of 404 when it should logically have been 200. That made me realize it should really be 404 in some situations, so the handler code had to be reworked to only return 200 for likely-good paths. We won't get to "known-good" paths until we support server side javascript (pre) rendering, at least not without a lot of overhead. For now this works.