Flesh out database connection with basic Plant type
There is now a Plant type Vertex record that can be defined and added. It is not terribly useful in itself yet, mostly just working out the kinks for how to approach db interop.

Also includes a basic unit test to demonstrate that the db connection and operations are working, at least against the in-memory db.